Promote base-minimal-test changes
Remove logstash integration for now, we are planning on moving to swift
for logs in vexxhost, which means we won't have logs indexes any more.

Also, we'd like to remove the dependency on sf-jobs, since they are
currently hosted in rdocloud, which results is job failures when the
cloud is having networking issues.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>