make pipeline precedence progressively higher
This patch changes the precedence settings for pipelines so that they
become progressively higher. Under high load, this is expected to have
the effect of prioritizing jobs needed to "finish" one patch
completely over jobs related to patches that are not as far along in
the approval process, which should prevent the post and gate pipeline
queues from backing up at the expense of making the check pipeline
queues deeper.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>